Architectures/3_PWA_Edge/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pydantic>=2.9.0
python-dotenv>=1.0.0
playwright>=1.44.0
